Update Python: Division.py

+++ Python: Division.py
@@ -3,4 +3,4 @@ if __name__ == '__main__':
     b = int(input())
 print(a//b)
 # print(int(a/b)
-print(float(a/b))
+print(a/b)
